rename all test files to use camelCase for names
